fix a signed/unsigned comparison in PR_UglyValueString
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@3538 d7cf8633-e32d-0410-b094-e92efae38249

+++ pr_edict.c
@@ -458,7 +458,7 @@ char *PR_UglyValueString (etype_t type, eval_t *val)
 		// (like newline, specifically) into escape codes,
 		// this fixes saving games from various mods
 		s = PR_GetString (val->string);
-		for (i = 0;i < sizeof (line) - 2 && *s;)
+		for (i = 0;i < (int)sizeof(line) - 2 && *s;)
 		{
 			if (*s == '\n')
 			{
